+++ src/anime_parsers_ru/parser_kodik_async.py
@@ -13,7 +13,7 @@ try:
     from . import errors # Импорт если библиотека установлена
     from .internal_tools import AsyncSession
 except ImportError:
-    import errors # Импорт если ббилиотека не установлена и файл лежит локально
+    import errors # Импорт если библиотека не установлена и файл лежит локально
     from internal_tools import AsyncSession
 
 class KodikParserAsync:
@@ -200,7 +200,7 @@ class KodikParserAsync:
             [
             {
                 "title": "Название",
-                "type": "тип мультимедия (anime, film, ...)",
+                "type": "тип мультимедиа (anime, film, ...)",
                 "year": "Год выпуска фильма",
                 "screenshots": [
                     "ссылки на скриншоты"
@@ -271,7 +271,7 @@ class KodikParserAsync:
         [
         {
             "title": "Название",
-            "type": "тип мультимедия (anime, film, ...)",
+            "type": "тип мультимедиа (anime, film, ...)",
             "year": "Год выпуска фильма",
             "screenshots": [
                 "ссылки на скриншоты"
@@ -312,7 +312,7 @@ class KodikParserAsync:
         [
         {
             "title": "Название",
-            "type": "тип мультимедия (anime, film, ...)",
+            "type": "тип мультимедиа (anime, film, ...)",
             "year": "Год выпуска фильма",
             "screenshots": [
                 "ссылки на скриншоты"
@@ -346,7 +346,7 @@ class KodikParserAsync:
 
     def _prettify_data(self, results: list[dict], only_anime: bool = False) -> list[dict]:
         """
-        Превращает полченные данные от запроса кодику в удобный вариант словаря
+        Превращает полученные данные от запроса кодику в удобный вариант словаря
 
         :results: список словарей (response['results'] в json'е от кодика)
         :only_anime: Возвращать только варианты аниме (тип anime или anime-serial) (по умолчанию False)
@@ -495,7 +495,7 @@ class KodikParserAsync:
                 'translations': self._generate_translations_dict(translations_div)
             }
         else:
-            raise errors.UnexpectedBehaviour('Ссылка на данные не была распознана как ссылка на сериал или фильм')
+            raise errors.UnexpectedBehavior('Ссылка на данные не была распознана как ссылка на сериал или фильм')
     
     def _is_serial(self, iframe_url: str) -> bool:
         return True if iframe_url[iframe_url.find(".info/")+6] == "s" else False